requirements.txt
redis
python-telegram-bot
rq